root/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
Deprecated==1.2.13
Django==4.0.3
django-countries==7.3.2
django-extensions==3.1.5
django-oauth-toolkit==1.7.0
django-oauth2-provider==0.2.6.1
django-phonenumber-field==6.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.1.1
mysqlclient==2.1.0
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.45
pycparser==2.21
pygraphviz==1.9
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shortuuid==1.0.8
simplejson==3.17.6
sqlparse==0.4.2
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
